Move event handlers to new handler classes
