+++ src/main/java/com/github/iunius118/tolaserblade/client/model/LaserBladeItemBakedModel.java
@@ -1,8 +1,8 @@
 package com.github.iunius118.tolaserblade.client.model;
 
-import com.github.iunius118.tolaserblade.ToLaserBladeConfig;
 import com.github.iunius118.tolaserblade.api.client.model.ILaserBladeModel;
 import com.github.iunius118.tolaserblade.client.model.laserblade.LaserBladeOBJModel;
+import com.github.iunius118.tolaserblade.config.ToLaserBladeConfig;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.Minecraft;
