+++ src/main/java/com/github/iunius118/tolaserblade/dispenser/DispenseLaserBladeBehavior.java
@@ -1,6 +1,6 @@
 package com.github.iunius118.tolaserblade.dispenser;
 
-import com.github.iunius118.tolaserblade.ToLaserBladeConfig;
+import com.github.iunius118.tolaserblade.config.ToLaserBladeConfig;
 import com.github.iunius118.tolaserblade.enchantment.ModEnchantments;
 import com.github.iunius118.tolaserblade.entity.LaserTrapEntity;
 import com.github.iunius118.tolaserblade.item.ModItems;
